modules/template/requirements.txt
flask==1.1.2
flask-cors==3.0.10
flask-restx==0.2.0
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-cloud-bigquery==2.7.0
google-cloud-error-reporting==1.1.0
google-cloud-logging==2.0.2
google-cloud-resource-manager==0.30.3
google-cloud-secret-manager==2.2.0
google-cloud-storage==1.35.0
google-cloud-tasks==2.1.0
gunicorn==20.0.4
psycopg2==2.8.6
